extracted editing & viewing, userState prop passed
